+++ src/Statistics.c
@@ -23,11 +23,11 @@
 struct opaque_online_average_struct
 {
   unsigned     n;
-  double       mean;
-  double       variance;
-  double       old_mean;
-  double       s;
-  double       old_s;
+  float        mean;
+  float        variance;
+  float        old_mean;
+  float        s;
+  float        old_s;
 };
 
 /*--------------------------------------------------------------------*/
@@ -69,32 +69,32 @@ int online_average_n(OnlineAverage* self)
 }
 
 /*--------------------------------------------------------------------*/
-double     online_average_mean(OnlineAverage* self)
+float      online_average_mean(OnlineAverage* self)
 {
   return self->mean;
 }
 
 /*--------------------------------------------------------------------*/
-double     online_average_variance(OnlineAverage* self)
+float      online_average_variance(OnlineAverage* self)
 {
   return self->variance;
 }
 
 /*--------------------------------------------------------------------*/
-double     online_average_std_dev(OnlineAverage* self)
+float      online_average_std_dev(OnlineAverage* self)
 {
-  return sqrt(self->variance);
+  return sqrtf(self->variance);
 }
 
 /*--------------------------------------------------------------------*/
-void online_average_update(OnlineAverage* self, double     x)
+void online_average_update(OnlineAverage* self, float     x)
 {
   ++self->n;
   
   if (self->n == 1)
     {
       self->old_mean = self->mean = x;
-      self->old_s    = 0.0;
+      self->old_s    = 0.0f;
     }
   else
     {
@@ -119,11 +119,11 @@ void online_average_update(OnlineAverage* self, double     x)
 struct opaque_moving_average_struct
 {
   unsigned       N;  // the nominal window size
-  double         mean;
-  double         variance;
-  double*        past_samples;
+  float          mean;
+  float          variance;
+  float*         past_samples;
   unsigned       current_sample_index;
-  double         s;
+  float          s;
   OnlineAverage* online;  //used until there are at least self->N samples
 };
 
@@ -185,27 +185,27 @@ int moving_average_n(MovingAverage* self)
 }
 
 /*--------------------------------------------------------------------*/
-double     moving_average_mean(MovingAverage* self)
+float      moving_average_mean(MovingAverage* self)
 {
   return self->mean;
 }
 
 /*--------------------------------------------------------------------*/
-double     moving_average_variance(MovingAverage* self)
+float      moving_average_variance(MovingAverage* self)
 {
   return self->variance;
 }
 
 /*--------------------------------------------------------------------*/
-double     moving_average_std_dev(MovingAverage* self)
+float      moving_average_std_dev(MovingAverage* self)
 {
-  return sqrt(self->variance);
+  return sqrtf(self->variance);
 }
 /*--------------------------------------------------------------------*/
-void moving_average_update(MovingAverage* self, double     x)
+void moving_average_update(MovingAverage* self, float     x)
 {
-  double     new_mean;
-  double     oldest_sample = self->past_samples[self->current_sample_index];
+  float     new_mean;
+  float     oldest_sample = self->past_samples[self->current_sample_index];
   
   self->past_samples[self->current_sample_index] = x;
   ++self->current_sample_index;
@@ -240,16 +240,16 @@ void moving_average_update(MovingAverage* self, double     x)
 struct opaque_online_regression_struct
 {
   unsigned     n         ;
-  double       a_mean    ;
-  double       b_mean    ;
-  double       a_variance;
-  double       b_variance;
-  double       covariance;
+  float        a_mean    ;
+  float        b_mean    ;
+  float        a_variance;
+  float        b_variance;
+  float        covariance;
   
   //slope, y-intercept, correlation
-  double       m         ;
-  double       b         ;
-  double       r_2       ;
+  float        m         ;
+  float        b         ;
+  float        r_2       ;
 };
 
 /*--------------------------------------------------------------------*/
@@ -294,40 +294,40 @@ int online_regression_n(OnlineRegression* self)
 }
 
 /*--------------------------------------------------------------------*/
-double     online_regression_covariance(OnlineRegression* self)
+float      online_regression_covariance(OnlineRegression* self)
 {
   return self->covariance;
 }
 
 /*--------------------------------------------------------------------*/
-double     online_regression_slope(OnlineRegression* self)
+float      online_regression_slope(OnlineRegression* self)
 {
   return self->m;
 }
 
 /*--------------------------------------------------------------------*/
-double     online_regression_y_intercept(OnlineRegression* self)
+float      online_regression_y_intercept(OnlineRegression* self)
 {
   return self->b;
 }
 
 /*--------------------------------------------------------------------*/
-double     online_regression_r_squared(OnlineRegression* self)
+float      online_regression_r_squared(OnlineRegression* self)
 {
   return self->r_2;
 }
 
 /*--------------------------------------------------------------------*/
-void online_regression_update(OnlineRegression* self, double     a_data /*x*/, double     b_data /*y*/)
+void online_regression_update(OnlineRegression* self, float     a_data /*x*/, float     b_data /*y*/)
 {
   ++self->n;
-  double     one_over_n = 1.0 / (double)self->n;
+  float     one_over_n = 1.0f / (float)self->n;
   
-  double     n_minus_1_over_n = (self->n-1) * one_over_n;
-  double     d_a = a_data - self->a_mean;
-  double     d_b = b_data - self->b_mean;
+  float     n_minus_1_over_n = (self->n-1) * one_over_n;
+  float     d_a = a_data - self->a_mean;
+  float     d_b = b_data - self->b_mean;
   
-  double     new_m, new_b;
+  float     new_m, new_b;
   self->a_variance += (n_minus_1_over_n * d_a * d_a - self->a_variance) * one_over_n;
   self->b_variance += (n_minus_1_over_n * d_b * d_b - self->b_variance) * one_over_n;
   self->covariance += (n_minus_1_over_n * d_a * d_b - self->covariance) * one_over_n;
@@ -358,17 +358,17 @@ void online_regression_update(OnlineRegression* self, double     a_data /*x*/, d
 ----------------------------------------------------------------------*/
 struct opaque_adaptive_threshold_struct
 {
-  double         smoothing; //coefficient 0 ~ 1
-  double         threshold; //standard deviations from filtered signal
-  double         min;
+  float          smoothing; //coefficient 0 ~ 1
+  float          threshold; //standard deviations from filtered signal
+  float          min;
   // step function (-1, 0, 1) that will hold its value while input is 
   // above threshold. update() only returns +1 or -1 at the moment
   // onset_signal transitions to +1 or -1;
-  double         onset_signal;
+  float          onset_signal;
   
   //(start with high variance to avoid spurious onsets on first iterations)
   MovingAverage* avg;
-  double         filtered_x;
+  float          filtered_x;
 };
 
 /*--------------------------------------------------------------------*/
@@ -389,8 +389,8 @@ AdaptiveThreshold* adaptive_threshold_new(unsigned N)
 /*--------------------------------------------------------------------*/
 void adaptive_threshold_init(AdaptiveThreshold* self)
 {
-  self->smoothing  = 0.5; //coefficient 0 ~ 1
-  self->threshold  = 3.5; //standard deviations from filtered signal
+  self->smoothing  = 0.5f; //coefficient 0 ~ 1
+  self->threshold  = 3.5f; //standard deviations from filtered signal
   self->min        = 0;
   adaptive_threshold_clear(self);
 }
@@ -415,77 +415,77 @@ AdaptiveThreshold* adaptive_threshold_destroy(AdaptiveThreshold* self)
 }
 
 /*--------------------------------------------------------------------*/
-double     adaptive_threshold_smoothing(AdaptiveThreshold* self)
+float      adaptive_threshold_smoothing(AdaptiveThreshold* self)
 {
   return self->smoothing;
 }
 
 /*--------------------------------------------------------------------*/
-void adaptive_threshold_set_smoothing(AdaptiveThreshold* self, double     coefficient)
+void adaptive_threshold_set_smoothing(AdaptiveThreshold* self, float     coefficient)
 {
   coefficient = (coefficient > 1) ? 1 : ((coefficient < 0) ? 0 : coefficient);
   self->smoothing = coefficient;
 }
 
 /*--------------------------------------------------------------------*/
-double     adaptive_threshold_threshold(AdaptiveThreshold* self)
+float      adaptive_threshold_threshold(AdaptiveThreshold* self)
 {
   return self->threshold;
 }
 
 /*--------------------------------------------------------------------*/
-double     adaptive_threshold_threshold_value(AdaptiveThreshold* self)
+float      adaptive_threshold_threshold_value(AdaptiveThreshold* self)
 {
-  double result = moving_average_std_dev(self->avg) * self->threshold;
+  float result = moving_average_std_dev(self->avg) * self->threshold;
   if(result < self->min)
     result = self->min;
   return result + moving_average_mean(self->avg);
 }
 
 /*--------------------------------------------------------------------*/
-void adaptive_threshold_set_threshold(AdaptiveThreshold* self, double     std_devs)
+void adaptive_threshold_set_threshold(AdaptiveThreshold* self, float     std_devs)
 {
   //if(std_devs < 0) std_devs = 0;
   self->threshold = std_devs;
 }
 
 /*--------------------------------------------------------------------*/
-double             adaptive_threshold_threshold_min    (AdaptiveThreshold* self)
+float              adaptive_threshold_threshold_min    (AdaptiveThreshold* self)
 {
   return self->min;
 }
 /*--------------------------------------------------------------------*/
-void               adaptive_threshold_set_threshold_min(AdaptiveThreshold* self, double min)
+void               adaptive_threshold_set_threshold_min(AdaptiveThreshold* self, float min)
 {
   //if(min < 0) min = 0;
   self->min = min;
 }
 
 /*--------------------------------------------------------------------*/
-double     adaptive_threshold_onset_signal(AdaptiveThreshold* self)
+float      adaptive_threshold_onset_signal(AdaptiveThreshold* self)
 {
   return self->onset_signal;
 }
 
 /*--------------------------------------------------------------------*/
-double             adaptive_threshold_mean          (AdaptiveThreshold* self)
+float              adaptive_threshold_mean          (AdaptiveThreshold* self)
 {
   return moving_average_mean(self->avg);
 }
 
 /*--------------------------------------------------------------------*/
-double     adaptive_threshold_update(AdaptiveThreshold* self, double     x)
+float      adaptive_threshold_update(AdaptiveThreshold* self, float     x)
 {
-  double     next   = 0;
-  double     result = 0;
+  float     next   = 0;
+  float     result = 0;
 
   self->filtered_x = (x * (1-self->smoothing)) + (self->filtered_x * self->smoothing);
   moving_average_update(self->avg, self->filtered_x);
 
-  double thresh = moving_average_std_dev(self->avg) * self->threshold;
+  float thresh = moving_average_std_dev(self->avg) * self->threshold;
   if(thresh < self->min) thresh = self->min;
 
-  if((fabs(x) - moving_average_mean(self->avg)) > thresh)
+  if((fabsf(x) - moving_average_mean(self->avg)) > thresh)
     next = (x > moving_average_mean(self->avg)) ? 1 : -1;
   
   //ignore the first few samples so we get a stable std dev
@@ -500,15 +500,15 @@ double     adaptive_threshold_update(AdaptiveThreshold* self, double     x)
 
 
 /*--------------------------------------------------------------------*/
-double     statistics_random_flat()
+float      statistics_random_flat()
 {
-  return random() / (double)RAND_MAX;
+  return random() / (float)RAND_MAX;
 }
 
 /*--------------------------------------------------------------------*/
-double     statistics_random_normal(double mean, double std_dev)
+float      statistics_random_normal(float mean, float std_dev)
 {
-  double     u, v, r;
+  float     u, v, r;
   
   do{
     u = 2 * statistics_random_flat() - 1;
@@ -516,20 +516,20 @@ double     statistics_random_normal(double mean, double std_dev)
     r = u*u + v*v;
   }while(r == 0 || r > 1);
 
-  double     c = sqrt(-2 * log(r) / r);
+  float     c = sqrtf(-2 * logf(r) / r);
   return mean + u * c * std_dev;
 }
 
 /*--------------------------------------------------------------------*/
-double     statistics_random_cauchy(double peak_location, double half_width_at_half_maximum)
+float      statistics_random_cauchy(float peak_location, float half_width_at_half_maximum)
 {
-  double result = statistics_random_flat();
-  result -= 0.5;
+  float result = statistics_random_flat();
+  result -= 0.5f;
   result *= M_PI;
-  //if(fabs(result) == M_PI)
+  //if(fabsf(result) == M_PI)
     //result = 0;
   //else
-    result = tan(result);
+    result = tanf(result);
   
   result *= half_width_at_half_maximum;
   result += peak_location;
